pyproject.toml
[tool.poetry]
name = "flask_address_book"
version = "0.1.0"
description = ""
authors = ["Vitaliy Lobko <vitaliy.lobko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
bcrypt = "^4.0.0"
python-dotenv = "^0.21.0"
marshmallow = "^3.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
